Use npm instead of bower to retrieve jQuery
The built version of jQuery that is in the bower package is considered
invalid by the addons.mozilla.org validator, because it does not match
the SHA256 checksum from the major CDNs. This difference is due to a
change where the CDN version of the file omits the sourcemap directive.